Add files via upload
In this report, I use the Amazon Food Review data set and conduct an analysis on it using the NLP method, as well as discuss the issues and problems with the data, as well as how I cleaned the data. With this data, I can also discuss interpretation and recommendations.